Reworking guns to be more accurate.
I hade erronously set the gun as the GSh-30, when it is in fact
the GSh-23. Also, the bullet velocity was wrong (corrected).

This also changes how ammo expenditure is calculated and is now
much better, now being time based vs. loop based.

Impact is now determined by the closest mp plane to the point of
impact (hard-coded limit of 35 meters). This change is also required
for the planned submodel-based rocket and bomb code.